\name{intrval-package}
\alias{intrval-package}
\docType{package}
\title{
\packageTitle{intrval}
}
\description{
\packageDescription{intrval}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{intrval}
\packageIndices{intrval}

Relational operators for value-to-interval comparisons:
\code{\link{\%[]\%}} and alikes.

Relational operators for interval-to-interval comparisons:
\code{\link{\%[o]\%}} and alikes.

Negated value matching: \code{\link{\%nin\%}}.
}
\author{
\packageAuthor{intrval}

Maintainer: \packageMaintainer{intrval}
}
\keyword{ package }
